--- src/actions/actions.ts
@@ -1,1 +1,0 @@
-export {};
