+++ src/components/TodoList/index.tsx
@@ -1,0 +1,13 @@
+import TodoListItem from "../TodoListItem";
+import styles from "./index.module.scss";
+
+const TodoList = () => {
+  return (
+    <>
+      <div>TodoList</div>
+      <TodoListItem />
+    </>
+  );
+};
+
+export default TodoList;
