--- src/components/addItemForm/index.js
@@ -1,3 +1,0 @@
-import AddItemForm from "./AddItemForm.tsx";
-
-export default AddItemForm;
