--- src/components/appHeader/index.js
@@ -1,3 +1,0 @@
-import AppHeader from "./AppHeader.tsx";
-
-export default AppHeader;
