--- src/components/todoList/TodoList.tsx
@@ -1,13 +1,0 @@
-import TodoListItem from "../todoListItem";
-import styles from './TodoList.module.scss';
-
-const TodoList = () => {
-  return (
-    <>
-    <div>TodoList</div>
-    <TodoListItem />
-    </>
-  )
-}
-
-export default TodoList;
