--- src/components/todoList/index.js
@@ -1,3 +1,0 @@
-import TodoList from "./TodoList.tsx";
-
-export default TodoList;
