--- src/components/todoListItem/TodoListItem.tsx
@@ -1,10 +1,0 @@
-import styles from "./TodoListItem.module.scss";
-
-const TodoListItem = () => {
-  
-    return (
-      <div>TodoListItem</div>
-    );
-}
-
-export default  TodoListItem;
