--- src/components/todoListItem/index.js
@@ -1,3 +1,0 @@
-import TodoListItem from "./TodoListItem.tsx";
-
-export default TodoListItem;
