+++ src/index.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from 'react-dom';
-import './index.css';
-import App from './components/app';
+import './index.module.scss';
+import App from './components/App';
 
 
 ReactDOM.render(
